[SPARK-20952] ParquetFileFormat should forward TaskContext to its forkjoinpool
See ticket [1] and the PR [2] @robert3005 opened on apache/spark. It
seems upstream wasn't convinced by our use-case.

We need this change because Spark reads parquet footers on a different
thread. Without this change, that thread doesn't inherit the
thread-local that stores the TaskContext, meaning we don't have access
to the properties stored inside that TaskContext

Our internal filesystem needs those properties in the task context and
fails to read footers without.

[1] https://issues.apache.org/jira/browse/SPARK-20952
[2] https://github.com/apache/spark/pull/18176

Co-authored-by: Robert Kruszewski <robertk@palantir.com>
Co-authored-by: Josh Casale <jcasale@palantir.com>
Co-authored-by: Will Raschkowski <wraschkowski@palantir.com>