Infer Pandas string columns in Arrow conversion on Python 2
When serializing a Pandas dataframe using Arrow under Python 2, Arrow
can't tell if string columns should be serialized as string type or as
binary (due to how Python 2 stores strings). The result is that Arrow
serializes string columns in Pandas dataframes to binary ones.

We can remove this when we discontinue support for Python 2.

See original PR [1] and follow-up for 'mixed' type columns [2].

[1] https://github.com/palantir/spark/pull/679
[2] https://github.com/palantir/spark/pull/702